Move files to v5 folder in source.
