+++ ecal/core/src/v5/service/ecal_service_server_impl.cpp
@@ -23,7 +23,7 @@
 
 #include <ecal/ecal_log.h>
 
-#include "ecal_service_server_v5_impl.h"
+#include "ecal_service_server_impl.h"
 
 namespace eCAL
 {
