+++ serialization/protobuf/protobuf/include/ecal/msg/protobuf/client.h
@@ -1,6 +1,6 @@
 /* ========================= eCAL LICENSE =================================
  *
- * Copyright (C) 2016 - 2024 Continental Corporation
+ * Copyright (C) 2016 - 2025 Continental Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,7 +25,7 @@
 #pragma once
 
 #include <ecal/ecal_deprecate.h>
-#include <ecal/ecal_client_v5.h>
+#include <ecal/v5/ecal_client.h>
 #include <ecal/msg/protobuf/ecal_proto_dyn.h>
 
 // protobuf includes
