+++ serialization/protobuf/protobuf/include/ecal/msg/protobuf/publisher.h
@@ -1,6 +1,6 @@
 /* ========================= eCAL LICENSE =================================
  *
- * Copyright (C) 2016 - 2024 Continental Corporation
+ * Copyright (C) 2016 - 2025 Continental Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,7 +27,7 @@
 #include <cstddef>
 #include <ecal/ecal_deprecate.h>
 #include <ecal/msg/protobuf/ecal_proto_hlp.h>
-#include <ecal/ecal_publisher_v5.h>
+#include <ecal/v5/ecal_publisher.h>
 
 // protobuf includes
 #ifdef _MSC_VER
